Schema Fields and Options name Field: This field is defined to store the name of the user.
type: Specifies that the data type for the name field is a string.
required: Indicates that the name field must be provided when creating a new user document.
email Field: This field is used to store the user's email address.

type: Specifies that the data type for the email field is a string.
unique: Ensures that each email address in the Users collection is unique, preventing duplicate email addresses.
password Field: This field is used to store the user's password.

type: Specifies that the data type for the password field is a string.
required: Indicates that the password field must be provided when creating a new user document.
role Field: This field is used to define the user's role, such as an admin or regular user.

type: Specifies that the data type for the role field is a number.
default: Sets the default value of the role field to 0 if no value is provided, typically indicating a regular user.
cart Field: This field is used to store the user's shopping cart items.

type: Specifies that the data type for the cart field is an array.
default: Sets the default value of the cart field to an empty array if no value is provided.