Creating and Exporting the Model mongoose.model: This function compiles the schema into a model. A model is a class with which we construct documents. In this case, the model is for the Users collection.
Model Name: The first argument, 'Users', is the name of the model. Mongoose automatically looks for the lowercase, plural version of your model name. For example, the model name Users will be converted to the users collection name in MongoDB.
Schema: The second argument is the userSchema which defines the structure of the documents in the collection.
module.exports: This exports the model so that it can be used in other parts of the application. By exporting the model, you can import it in other files to interact with the Users collection, such as creating, reading, updating, or deleting user documents.