app.use(express.json()) This line uses the express.json() middleware:
JSON Parsing Middleware: The express.json() function is a built-in middleware in Express that parses incoming requests with JSON payloads. It is particularly useful for APIs that accept JSON-formatted data.
Usage: By calling app.use(express.json()), the middleware is applied to all incoming requests. It parses the request body and makes the JSON data available in req.body.
4. app.use(cookieParser())
This line uses the cookie-parser middleware:

Cookie Parsing Middleware: The cookieParser() function initializes the cookie-parser middleware. This middleware parses cookies attached to the client request object.
Usage: By calling app.use(cookieParser()), the middleware is applied to all incoming requests. It populates req.cookies with an object containing the cookies sent by the client.