1. require('dotenv').config(); This line imports and configures the dotenv package:
Importing dotenv: The require('dotenv') function imports the dotenv package, which is used to load environment variables from a .env file.
Configuring dotenv: Calling .config() on the imported dotenv package loads the environment variables from the .env file into process.env. This allows the application to access configuration settings and secrets stored in the .env file throughout the code.
2. const cookieParser = require('cookie-parser')
This line imports the cookie-parser middleware:

Importing cookie-parser: The require('cookie-parser') function imports the cookie-parser package, which is used to parse cookies attached to client requests. This middleware creates a req.cookies object, where cookies sent by the client are stored and accessible in the request object.