Password Validation:
If a user with the provided email exists, the function proceeds to validate the provided password against the hashed password stored in the user record.
It uses bcrypt.compare() to compare the hashed password retrieved from the database (user.password) with the provided plaintext password (password).
If the passwords match, bcrypt.compare() returns true, indicating a successful password match. If not, it returns false, indicating an incorrect password.
If the password doesn't match, the function returns a 400 status response with the message "Incorrect Password".