Token Generation:
If both the email and password are valid, the function generates a new access token and a refresh token.
The access token is created using the createAccessToken function, passing the user's ID (user._id) as the payload.
Similarly, the refresh token is created using the createRefreshToken function, also passing the user's ID as the payload.