Cookie Setting:
The function sets the refresh token as an HTTP-only cookie named "refreshtoken". This is done using res.cookie(), which is provided by Express.
The cookie is set to be HTTP-only, meaning it cannot be accessed via client-side JavaScript, enhancing security.
It specifies the path of the cookie as '/user/refresh_token', indicating that this cookie should only be sent to requests under the '/user/refresh_token' path.