Clearing Refresh Token Cookie:
The logout function starts by clearing the refresh token cookie from the client's browser.
It uses res.clearCookie() provided by Express to remove the cookie named 'refreshtoken'.
The path option specifies the path of the cookie to be cleared. In this case, it's set to '/user/refresh_token' to match the path where the cookie was originally set.