Token Verification:
After retrieving the token, the middleware checks if the token exists. If it doesn't, indicating that the user is not authenticated, it returns a 400 status response with the message "Invalid Authentication".
If the token exists, the middleware proceeds to verify its authenticity using jwt.verify().
It calls jwt.verify() with the token, the secret key (process.env.ACCESS_TOKEN_SECRET), and a callback function to handle the verification result.
If the token is successfully verified (i.e., the signature matches and it hasn't expired), the callback function is called with the decoded payload (user) as an argument.
If there's an error during verification (e.g., invalid token or signature), the callback function is called with the error object as the first argument.