Request User Population:
If the token is successfully verified, the decoded payload (user) is assigned to req.user. This allows subsequent middleware functions or route handlers to access the authenticated user's information.
The middleware then calls next() to pass control to the next middleware function or route handler in the request handling chain.