Error Handling:
The middleware is wrapped in a try-catch block to catch any potential errors that may occur during its execution.
If an error occurs, such as an error in retrieving the token or during token verification, it returns a 500 status response with the error message.
In summary, the auth middleware function intercepts incoming requests, retrieves and verifies JWTs provided in the 'Authorization' header, and populates the req.user object with the authenticated user's information if the token is valid. It then passes control to the next middleware function or route handler.