Password Exclusion:
Once the user document is retrieved, the function removes the password field from the user object before sending it back in the response.
This is achieved using .select('-password'), which excludes the 'password' field from the user object.
Excluding the password field enhances security by ensuring that sensitive information is not exposed to the client.