User Query:
Inside the middleware function, it queries the database to find the user associated with the authenticated user's ID (req.user.id).
It uses Users.findOne() to search for a user document in the database that matches the provided query criteria (in this case, the user's ID).
The await keyword is used to wait for the database query to complete asynchronously, allowing the function to proceed once the user document is retrieved.