Admin Role Check:
After retrieving the user document, the middleware checks if the user has an admin role.
If the user's role field is equal to 0, it indicates that the user is not an admin.
In such a case, the middleware returns a 400 status response with the message "Admin Resources Access Denied", indicating that access to admin resources is denied for non-admin users.