Passing Control to Next Middleware/Route Handler:
If the user has an admin role (i.e., role is not equal to 0), the middleware calls the next() function to pass control to the next middleware function or route handler in the request handling chain.
This allows subsequent middleware functions or route handlers to be invoked, enabling further processing of the request.