Schema Fields:
Within the schema definition, there is a single field named name.
name is defined as an object with properties specifying the field type, whether it's required, and additional validation rules.
In this case:
type: String specifies that the field should store string values.
required: true indicates that the field is required and must have a value for documents to be considered valid.
trim: true ensures that any leading or trailing whitespace is trimmed from the string value.
unique: true specifies that the values in this field must be unique across all documents in the collection.