Schema Options:
After defining the fields, an optional second argument is provided to mongoose.Schema(), which contains schema options.
In this case, timestamps: true is specified as an option.
timestamps is a Mongoose schema option that automatically adds two fields (createdAt and updatedAt) to the schema. These fields track the creation and last modification times of documents.