getCategories Function:
getCategories is an asynchronous function that handles the retrieval of all categories.
It uses Category.find() to fetch all documents from the 'Category' collection.
Upon successful retrieval, it sends the categories as a JSON response.
If an error occurs during the process, it returns a 500 status code along with an error message.