Schema Fields:
The schema defines several fields for a product document:
product_id: A unique identifier for the product, specified as a string.
title: The title of the product, specified as a string.
price: The price of the product, specified as a number.
description: A description of the product, specified as a string.
content: Additional content or details about the product, specified as a string.
images: Images related to the product, specified as an object.
category: The category to which the product belongs, specified as a string.
checked: A boolean value indicating whether the product is checked or not, with a default value of false.
sold: The number of units sold for the product, with a default value of 0.