Exporting the Product Model:
Finally, mongoose.model("Products", productSchema) creates a model for the 'Products' collection based on the defined schema.
This model is exported for use in other parts of the application to perform CRUD operations on documents in the 'Products' collection.
In summary, the code defines a Mongoose schema for documents in the 'Products' collection with various fields representing product attributes. The schema includes options to automatically track timestamps. Finally, a model is created based on the schema and exported for use in other parts of the application.