Constructor:
constructor(query, queryString) { ... }: The constructor method is a special method that is automatically called when you create a new instance of the APIfeatures class. It accepts two parameters: query and queryString.
query: This parameter represents the MongoDB query object. It typically contains conditions and options for querying the database.
queryString: This parameter represents the query string parameters received from the HTTP request. In Express.js, it's commonly accessed via req.query and contains parameters for filtering, sorting, pagination, etc.