Filtering Method:
filtering() { ... }: This method is responsible for filtering the MongoDB query based on the query string parameters.
It creates a copy of the queryString object using the spread operator ({...this.queryString}) to avoid directly modifying the original object.
It removes excluded fields (page, sort, limit) from the queryObj.
It converts the queryObj into a JSON string (queryStr) and performs string replacement using a regular expression to replace specific operators (gte, gt, lt, lte, regex) with MongoDB query operators.
It updates the query property of the APIfeatures instance by applying the filtered query conditions.