Sorting Method:
sorting() { ... }: This method is responsible for sorting the MongoDB query based on the sort query string parameter.
If the sort parameter exists in the queryString, it splits the value by commas, joins the resulting array elements with a space, and assigns the sorted field and order to the sortBy variable.
It updates the query property of the APIfeatures instance to sort the results accordingly. If sort parameter is not provided, it defaults to sorting by -createdAt.