updateProduct:
This method handles the update of a product by its ID.
It extracts the updated product details from the request body (req.body).
It checks if images are provided and returns a 500 status with an error message if not.
It uses the findOneAndUpdate() method provided by Mongoose to find the product by its ID and update its details with the provided data.
Upon successful update, it sends a JSON response confirming the update of the product.
Overall, the productCtrl object encapsulates the logic for handling CRUD (Create, Read, Update, Delete) operations related to products in the application. Each method corresponds to a specific CRUD operation and interacts with the database using Mongoose methods. Additionally, it handles error cases and sends appropriate responses to the client.