UserAPI Component State Declarations:
isLogged: A state variable that indicates whether the user is logged in. Initially set to false.
setIsLogged: A function to update the isLogged state.
isAdmin: A state variable that indicates whether the logged-in user is an admin. Initially set to false.
setIsAdmin: A function to update the isAdmin state.
cart: A state variable that holds the array of items in the user's cart. Initially set to an empty array.
setCart: A function to update the cart state.